distance_sensor.ino
A reverse parking mechanism using Arduino which alerts drivers of any obstacles behind them while parking. The Ultrasonic sensor sends out a high-frequency sound pulse which bounces off any nearby object and returns to the sensor.  

The time taken for this process is measured by the Arduino which then calculates the distance between the car and the object. The system then alerts the driver of any obstacles behind them by sounding an alarm.